test(contactout): consolidate coverage into shared suites
