+++ tests/test_capacity_collectors.py
@@ -6,6 +6,9 @@ Each test mocks the upstream API response and verifies that the collector return
 
 from __future__ import annotations
 
+from datetime import timedelta
+from treg.domain.capacity import policy, sweep
+from treg.timeutil import utcnow_naive
 from treg.domain.capacity import collectors
 import httpx
 import pytest
@@ -177,3 +180,55 @@ def test_implemented_collectors_are_registered_and_do_not_overlap_absent_list():
         assert provider not in collectors.NO_BALANCE_API
     overlap = set(collectors.BALANCE_ROUTES.keys()) & set(collectors.NO_BALANCE_API.keys())
     assert not overlap, f"Providers in both maps: {overlap}"
+
+
+# ---- ContactOut ----
+
+async def test_contactout_pool_stats_are_informational_even_when_empty(contactout_platform):
+    for extra in ({}, {"remaining": -5, "phone_remaining": 0, "search_remaining": 20}):
+        usage = {
+            "count": 10,
+            "quota": 0,
+            "phone_count": 20,
+            "phone_quota": 0,
+            "search_count": 30,
+            "search_quota": 20,
+        } | extra
+        async with httpx.AsyncClient(
+            transport=httpx.MockTransport(
+                lambda r: httpx.Response(200, json={"status_code": 200, "usage": usage})
+            )
+        ) as c:
+            row = await collectors.provider_balance("contactout", c)
+        snap = sweep.snapshot_from("contactout", row)
+        assert not snap.error and snap.remaining is None
+        assert "email: used=10, quota=0" in snap.note
+        state = policy.latest_state(
+            policy.default_policy("contactout", has_key=True), snap
+        )
+        assert state.confidence == "informational" and state.exhausted_until is None
+        old = policy.latest_state(
+            policy.default_policy("contactout", has_key=True),
+            snap,
+            now=utcnow_naive() + timedelta(hours=7),
+        )
+        assert old.health == "stale" and old.exhausted_until is None
+
+
+@pytest.mark.parametrize(
+    "payload",
+    [
+        {},
+        {"status_code": 401},
+        {"status_code": 200, "usage": {}},
+        {"status_code": 200, "usage": {"count": True, "quota": 1}},
+    ],
+)
+async def test_contactout_bad_stats_are_not_valid_observations(contactout_platform, payload):
+    async with httpx.AsyncClient(
+        transport=httpx.MockTransport(lambda r: httpx.Response(200, json=payload))
+    ) as c:
+        row = await collectors.provider_balance("contactout", c)
+    assert row["value"] is None and not row.get("informational")
+    assert "PLATFORM-TEST" not in str(row)
+    assert sweep.snapshot_from("contactout", row).error
